requirements.txt
scikit-learn
pandas
numpy==1.19.3
flask
flask-cors
matplotlib
seaborn
scipy==0.19.1
wordcloud
plotly
torch==2.1.2
torchtext==0.16.2
torchvision
transformers
Pillow
streamlit
nbformat
nltk
datasets
streamlit-extras
prometheus-client
grafana-api
docker
jupyter
optuna
yfinance
tqdm
joblib
opencv-python
face-recognition
python-dotenv
surprise
# auto-sklearn
